+++ R/model_parameters_default.R
@@ -10,6 +10,7 @@
 #' @param exponentiate Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link.
 #' @param robust Logical, if \code{TRUE}, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like \code{vcov_estimation} or \code{vcov_type} are passed down to other methods, see \code{\link[=standard_error_robust]{standard_error_robust()}} for details.
 #' @param component Model component for which parameters should be shown. May be one of \code{"conditional"}, \code{"precision"} (\pkg{betareg}), \code{"scale"} (\pkg{ordinal}), \code{"extra"} (\pkg{glmx}) or \code{"all"}.
+#' @param p_adjust Character vector, if not \code{NULL}, indicates the method to adjust p-values. See \code{\link[stats]{p.adjust}} for details.
 #' @param ... Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.
 #'
 #' @seealso \code{\link[=standardize_names]{standardize_names()}} to rename
@@ -39,7 +40,7 @@
 #' model_parameters(model, exponentiate = TRUE)
 #' @return A data frame of indices related to the model's parameters.
 #' @export
-model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, standardize = NULL, exponentiate = FALSE, robust = FALSE, ...) {
+model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, standardize = NULL, exponentiate = FALSE, robust = FALSE, p_adjust = NULL, ...) {
   out <- .model_parameters_generic(
     model = model,
     ci = ci,
@@ -50,6 +51,7 @@ model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterati
     exponentiate = exponentiate,
     effects = "fixed",
     robust = robust,
+    p_adjust = p_adjust,
     ...
   )
 
@@ -59,7 +61,7 @@ model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterati
 
 
 
-.model_parameters_generic <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, merge_by = "Parameter", standardize = NULL, exponentiate = FALSE, effects = "fixed", robust = FALSE, df_method = NULL, ...) {
+.model_parameters_generic <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, merge_by = "Parameter", standardize = NULL, exponentiate = FALSE, effects = "fixed", robust = FALSE, df_method = NULL, p_adjust = NULL, ...) {
   # to avoid "match multiple argument error", check if "component" was
   # already used as argument and passed via "...".
   mc <- match.call()
@@ -70,9 +72,9 @@ model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterati
     parameters <- bootstrap_parameters(model, iterations = iterations, ci = ci, ...)
   } else {
     parameters <- if (is.null(comp_argument)) {
-      .extract_parameters_generic(model, ci = ci, component = "conditional", merge_by = merge_by, standardize = standardize, effects = effects, robust = robust, df_method = df_method, ...)
+      .extract_parameters_generic(model, ci = ci, component = "conditional", merge_by = merge_by, standardize = standardize, effects = effects, robust = robust, df_method = df_method, p_adjust = p_adjust, ...)
     } else {
-      .extract_parameters_generic(model, ci = ci, merge_by = merge_by, standardize = standardize, effects = effects, robust = robust, df_method = df_method, ...)
+      .extract_parameters_generic(model, ci = ci, merge_by = merge_by, standardize = standardize, effects = effects, robust = robust, df_method = df_method, p_adjust = p_adjust, ...)
     }
   }
 
@@ -91,7 +93,7 @@ model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterati
 
 #' @rdname model_parameters.default
 #' @export
-model_parameters.betareg <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("conditional", "precision", "all"), standardize = NULL, exponentiate = FALSE, ...) {
+model_parameters.betareg <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("conditional", "precision", "all"), standardize = NULL, exponentiate = FALSE, p_adjust = NULL, ...) {
   component <- match.arg(component)
   if (component == "all") {
     merge_by <- c("Parameter", "Component")
@@ -111,6 +113,7 @@ model_parameters.betareg <- function(model, ci = .95, bootstrap = FALSE, iterati
     standardize = standardize,
     exponentiate = exponentiate,
     robust = FALSE,
+    p_adjust = p_adjust,
     ...
   )
 
@@ -122,7 +125,7 @@ model_parameters.betareg <- function(model, ci = .95, bootstrap = FALSE, iterati
 
 #' @rdname model_parameters.default
 #' @export
-model_parameters.clm2 <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("all", "conditional", "scale"), standardize = NULL, exponentiate = FALSE, ...) {
+model_parameters.clm2 <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("all", "conditional", "scale"), standardize = NULL, exponentiate = FALSE, p_adjust = NULL, ...) {
   component <- match.arg(component)
   if (component == "all") {
     merge_by <- c("Parameter", "Component")
@@ -142,6 +145,7 @@ model_parameters.clm2 <- function(model, ci = .95, bootstrap = FALSE, iterations
     standardize = standardize,
     exponentiate = exponentiate,
     robust = FALSE,
+    p_adjust = p_adjust,
     ...
   )
 
@@ -156,7 +160,7 @@ model_parameters.clmm2 <- model_parameters.clm2
 
 #' @rdname model_parameters.default
 #' @export
-model_parameters.glmx <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("all", "conditional", "extra"), standardize = NULL, exponentiate = FALSE, ...) {
+model_parameters.glmx <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("all", "conditional", "extra"), standardize = NULL, exponentiate = FALSE, p_adjust = NULL, ...) {
   component <- match.arg(component)
   if (component == "all") {
     merge_by <- c("Parameter", "Component")
@@ -174,6 +178,7 @@ model_parameters.glmx <- function(model, ci = .95, bootstrap = FALSE, iterations
     standardize = standardize,
     exponentiate = exponentiate,
     robust = FALSE,
+    p_adjust = p_adjust,
     ...
   )